+++ testing/snapshot_test.ts
@@ -9,9 +9,8 @@ import {
   assertStringIncludes,
   fail,
 } from "@std/assert";
-import { assertSnapshot, createAssertSnapshot } from "./snapshot.ts";
+import { assertSnapshot, createAssertSnapshot, serialize } from "./snapshot.ts";
 import { ensureDir } from "../fs/ensure_dir.ts";
-import { serialize } from "./_snapshot_utils.ts";
 
 const SNAPSHOT_MODULE_URL = toFileUrl(join(
   dirname(fromFileUrl(import.meta.url)),
